feat: 특정 역대콘 상세 조회 api 추가

+++ libs/soongan-persistence/src/main/kotlin/com/soongan/soonganbackend/soonganpersistence/storage/weeklyContestFinal/WeeklyContestFinalEntity.kt
@@ -44,7 +44,7 @@ data class WeeklyContestFinalEntity(
     val ranking: Int,
 
     @Column(name = "score")
-    val score: BigDecimal
+    val score: Int
 ) {
 
     @CreatedDate
